pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ca-boil-experiments"
version = "0.1.0"
description = "Reproducible experiments for CA-BOIL vs BOIL on Fashion-MNIST"
authors = ["AI Researcher <research@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.1.0"
torchvision = "^0.16.0"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
wandb = "^0.16.0"
optuna = "^3.4.0"
scikit-learn = "^1.3.2"
scipy = "^1.11.3"
numpy = "^1.26.2"
pandas = "^2.1.3"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
tqdm = "^4.66.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"